DeclareOrder (ClientMessaging)

+++ src/main/java/example/Order.kt
@@ -4,15 +4,12 @@ class Order(
         val type: PackageType,
         val client : Client,
         val endTime : Long,
-        val price : Int,
-        val fine : Int
+        val price : Double,
+        val fine : Double
 ){
 
     var isPickedUp = false
     var isDelivered = false
     var warehouse: Warehouse? = null
     var drone: Drone? = null
-
-    //val pickupLocation : Point
-    //  get() = client.position!!.get()
 }
